(Logical change 1.2504)

+++ include/aal/bitops.h
@@ -9,15 +9,15 @@
 #define _ROUND_UP(x, n) (((x)+(n) - 1u) & ~((n) - 1u))
 #define ROUND_UP(x) _ROUND_UP(x, 8ll)
 
-extern inline int aal_set_bit (unsigned long nr, void *addr);
-extern inline int aal_clear_bit (unsigned long nr, void *addr);
-extern inline int aal_test_bit(unsigned long nr, const void *addr);
+extern inline int aal_set_bit (unsigned long long nr, void *addr);
+extern inline int aal_clear_bit (unsigned long long nr, void *addr);
+extern inline int aal_test_bit(unsigned long long nr, const void *addr);
 
-extern inline int aal_find_first_zero_bit (const void *vaddr, 
-    unsigned long size);
+extern inline unsigned long long aal_find_first_zero_bit (const void *vaddr, 
+    unsigned long long size);
 
-extern inline int aal_find_next_zero_bit (const void *vaddr, 
-    unsigned long size, unsigned long offset);
+extern inline unsigned long long aal_find_next_zero_bit (const void *vaddr, 
+    unsigned long long size, unsigned long long offset);
 
 #endif
